requirements.txt
requests>=2.31.0
pyyaml>=6.0.0
flask>=3.0.0
python-dotenv>=1.0.0
py-clob-client>=0.12.0
